+++ eightyhd.py
@@ -1,7 +1,8 @@
 from flask import Flask, render_template, request, redirect, url_for
 from flask_sqlalchemy import SQLAlchemy
-from datetime import datetime
+from datetime import datetime, timedelta
 from dateutil import parser
+from sqlalchemy import func, Date
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///eightyhd.db'
@@ -26,9 +27,10 @@ class Completed(db.Model):
 @app.route('/')
 def index():
     # Show all bounties
-    bounties = Bounties.query.filter(Bounties.status == 'pending').all()
+    recurring_bounties = Bounties.query.filter_by(status='pending', task_type='recurring').all()
+    single_bounties = Bounties.query.filter_by(status='pending', task_type='single').all()
     total_points = get_total_points()
-    return render_template('index.html', bounties=bounties, total_points=total_points)
+    return render_template('index.html', recurring_bounties=recurring_bounties, single_bounties=single_bounties, total_points=total_points)
 
 @app.route('/add', methods=['POST'])
 def add_task():
@@ -70,6 +72,29 @@ def add_completed():
     db.session.commit()
     return redirect(url_for('index'))
 
+@app.route('/daily-points')
+def daily_points():
+    today = datetime.utcnow().date()
+    seven_days_ago = today - timedelta(days=7)
+
+    # Query to sum points for the last 7 days, grouped by day
+    points_per_day = db.session.query(
+        func.date(Completed.complete_date).label('day'),
+        func.sum(Completed.reward).label('total_points')
+    ).filter(Completed.complete_date >= seven_days_ago).group_by(func.date(Completed.complete_date)).all()
+
+    # Create a dictionary to fill in missing days with zero points
+    days_points = {today - timedelta(days=i): 0 for i in range(7)}
+    for day in points_per_day:
+        # Ensure that `day.day` is a datetime.date object
+        days_points[day.day] = day.total_points
+
+    # Ensure all dictionary keys are datetime.date objects before sorting
+    #sorted_days_points = sorted(days_points.items())
+    sorted_days_points = sorted((datetime.strptime(key, '%Y-%m-%d').date() if isinstance(key, str) else key, value) for key, value in days_points.items())
+
+    return render_template('daily_points.html', points=sorted_days_points)
+
 def get_total_points():
     total_points = db.session.query(db.func.sum(Completed.reward)).scalar() or 0
     return total_points
@@ -77,4 +102,4 @@ def get_total_points():
 if __name__ == "__main__":
     with app.app_context():
         db.create_all()
-    app.run(debug=True)
+    app.run(host='0.0.0.0', debug=True, port=5009)